refactor modeling pipeline into multiple subscripts
